Fix tests and integration functionality.
